# ci/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VideoOrganizer API",
    "description": "Self-hosted media library manager \u2014 videos, tags, custom properties, background workers (thumbnail sprites \u002B Md5 dedup), and bulk import.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/videos": {
      "get": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "GetVideos",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "CreateVideo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Video"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/videos/{id}": {
      "get": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "GetVideoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "UpdateVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "DeleteVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/video-sets": {
      "get": {
        "tags": [
          "VideoSets"
        ],
        "operationId": "ListVideoSets",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "VideoSets"
        ],
        "operationId": "CreateVideoSet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoSet"
                }
              }
            }
          }
        }
      }
    },
    "/api/video-sets/{id}": {
      "put": {
        "tags": [
          "VideoSets"
        ],
        "operationId": "UpdateVideoSet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoSet"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "VideoSets"
        ],
        "operationId": "DeleteVideoSet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "force",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/video-sets/{id}/orphan-count": {
      "get": {
        "tags": [
          "VideoSets"
        ],
        "operationId": "GetVideoSetOrphanCount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/video-sets/{id}/re-root/preview": {
      "post": {
        "tags": [
          "VideoSets"
        ],
        "operationId": "ReRootVideoSetPreview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReRootPreview"
                }
              }
            }
          }
        }
      }
    },
    "/api/video-sets/{id}/re-root": {
      "post": {
        "tags": [
          "VideoSets"
        ],
        "operationId": "ReRootVideoSet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CameraTypes": {
        "enum": [
          "unknown",
          "cellPhone",
          "hiddenCamera",
          "camcorder",
          "professionalCamera",
          "notChecked"
        ],
        "type": "string"
      },
...
      "ChapterMarkerDto": {
        "required": [
          "offset",
          "comment"
        ],
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DomainVideoCodec": {
        "enum": [
          "h264",
          "h265",
          "hevc",
          "notChecked",
          "other"
        ],
        "type": "string"
      },
...
      "ReRootPreview": {
        "required": [
          "totalAffected",
          "sampled",
          "found",
          "missing",
          "newBaseExists",
          "examples"
        ],
        "type": "object",
        "properties": {
          "totalAffected": {
            "type": "integer",
            "format": "int32"
          },
          "sampled": {
            "type": "integer",
            "format": "int32"
          },
          "found": {
            "type": "integer",
            "format": "int32"
          },
          "missing": {
            "type": "integer",
            "format": "int32"
          },
          "newBaseExists": {
            "type": "boolean"
          },
          "examples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReRootPreviewItem"
            }
          }
        }
      },
...
      "ReRootRequest": {
        "required": [
          "newPath"
        ],
        "type": "object",
        "properties": {
          "newPath": {
            "type": "string"
          }
        }
      },
...
      "UpdateVideoRequest": {
        "required": [
          "fileName",
          "ingestDate",
          "cameraType",
          "videoQuality",
          "watchCount",
          "notes",
          "needsReview",
          "isFavorite",
          "clipStartSeconds",
          "clipEndSeconds",
          "chapterMarkers",
          "videoBlocks",
          "tagIds",
          "properties"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/CameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/VideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "chapterMarkers": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ChapterMarkerDto"
            }
          },
          "videoBlocks": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/VideoBlockDto"
            }
          },
          "tagIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "properties": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PropertyValueWrite"
            }
          }
        }
      },
      "Video": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "md5": {
            "type": [
              "null",
              "string"
            ]
          },
          "md5Failed": {
            "type": "boolean"
          },
          "md5FailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsFailed": {
            "type": "boolean"
          },
          "thumbnailsFailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsGenerated": {
            "type": "boolean"
          },
          "importJobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "pattern": "^-?(\\d\u002B\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "videoDimensionFormat": {
            "$ref": "#/components/schemas/DomainVideoDimensionFormat"
          },
          "videoCodec": {
            "$ref": "#/components/schemas/DomainVideoCodec"
          },
          "bitrate": {
            "type": "integer",
            "format": "int64"
          },
          "frameRate": {
            "type": "number",
            "format": "double"
          },
          "pixelFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "ratio": {
            "type": [
              "null",
              "string"
            ]
          },
          "creationTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "videoStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/DomainCameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/DomainVideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "playbackIssue": {
            "type": "boolean"
          },
          "markedForDeletion": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "parentVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isClip": {
            "type": "boolean"
          },
          "isExportedClip": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "clipExported": {
            "type": "boolean"
          },
          "exportedToVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "ocrScannedThroughSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "chapterMarkers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterMarker"
            }
          },
          "videoBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoBlock"
            }
          },
          "ocrTextLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcrTextLine"
            }
          },
          "videoTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoTag"
            }
          },
          "propertyValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoPropertyValue"
            }
          }
        }
      },
...
      "VideoCodec": {
        "enum": [
          "h264",
          "h265",
          "hevc",
          "notChecked",
          "other"
        ],
        "type": "string"
      },
      "VideoDimensionFormat": {
        "enum": [
          "uhd8k",
          "uhd4K",
          "hd1080p",
          "hd720p",
          "sd576p4x3",
          "sd576p16x9",
          "sd480p4x3",
          "sd480p16x9",
          "verticalUHD4k",
          "vertical1080p",
          "vertical720p",
          "nonStandard"
        ],
        "type": "string"
      },
      "VideoDto": {
        "required": [
          "id",
          "fileName",
          "filePath",
          "md5",
          "md5Failed",
          "md5FailedError",
          "thumbnailsFailed",
          "thumbnailsFailedError",
          "thumbnailsGenerated",
          "importJobId",
          "fileSize",
          "duration",
          "height",
          "width",
          "videoDimensionFormat",
          "videoCodec",
          "bitrate",
          "frameRate",
          "pixelFormat",
          "ratio",
          "creationTime",
          "videoStreamCount",
          "audioStreamCount",
          "ingestDate",
          "cameraType",
          "videoQuality",
          "watchCount",
          "notes",
          "needsReview",
          "playbackIssue",
          "markedForDeletion",
          "isFavorite",
          "parentVideoId",
          "clipStartSeconds",
          "clipEndSeconds",
          "isClip",
          "isExportedClip",
          "isEdited",
          "chapterMarkers",
          "videoBlocks",
          "tags",
          "properties"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "md5": {
            "type": [
              "null",
              "string"
            ]
          },
          "md5Failed": {
            "type": "boolean"
          },
          "md5FailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsFailed": {
            "type": "boolean"
          },
          "thumbnailsFailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsGenerated": {
            "type": "boolean"
          },
          "importJobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "pattern": "^-?(\\d\u002B\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "videoDimensionFormat": {
            "$ref": "#/components/schemas/VideoDimensionFormat"
          },
          "videoCodec": {
            "$ref": "#/components/schemas/VideoCodec"
          },
          "bitrate": {
            "type": "integer",
            "format": "int64"
          },
          "frameRate": {
            "type": "number",
            "format": "double"
          },
          "pixelFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "ratio": {
            "type": [
              "null",
              "string"
            ]
          },
          "creationTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "videoStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/CameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/VideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "playbackIssue": {
            "type": "boolean"
          },
          "markedForDeletion": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "parentVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isClip": {
            "type": "boolean"
          },
          "isExportedClip": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "chapterMarkers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterMarkerDto"
            }
          },
          "videoBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoBlockDto"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoTagSummaryDto"
            }
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyValueDto"
            }
          }
        }
      },
...
      "VideoQuality": {
        "enum": [
          "singleCamera",
          "multipleCameras",
          "lowQuality",
          "notChecked"
        ],
        "type": "string"
      },
      "VideoSet": {
        "required": [
          "name",
          "path"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}